pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "archist"
version = "1.0.0"
description = "A Python package"
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Brandon Wong"},
]
keywords = []
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "keystone-engine",
    "capstone",
    "unicorn",
    "qiling",
    "lief",
    "pyelftools",
    "pyghidra"
]

[project.optional-dependencies]
dev = [
    "Jinja2",
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "pre-commit",
    "ghidra-stubs",
]

[tool.hatch.build.targets.wheel]
packages = ["src/archist"]

[tool.ruff.lint]
preview = true
extend-select = ["E303", "W391"]

[tool.ruff.format]
exclude = ["src/archist/__init__.py", "src/archist/modes.py"]
